{{cookiecutter.project_slug}}/app/requirements.txt
connexion[swagger-ui]
flask
flask-migrate
flask-sqlalchemy
gunicorn
mysqlclient
pytest
sentry-sdk[flask]
